backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.1
paho-mqtt==2.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
python-dateutil==2.8.2
APScheduler==3.10.4
